Moved all app settings into the SavedSettings class.
Still need to rewrite the cache check code and the app startup code.

Also, seem to have found the cache size bug that caused one user to not be able to cache more than halfway. It was setting the minCacheSize when it should have been setting maxCacheSize in the settings tab. May not be it, but could be.

THIS COMMIT DOES NOT BUILT YET, MORE ERRORS TO FIX