Removed built frameworks from gitignore
A big issue with the old repository setup was all the submodules and how difficult it was to just check out an old commit and build it. Carthage is better in this regard, but as it’s only a few frameworks, I think it’s  better to just check eveything in.